Updated palworld mod manager script to correctly sync to client and server Palworld dirs

+++ scripts/palworld-mods.py
@@ -4,6 +4,20 @@ import argparse
 from pathlib import Path
 
 
+def read_env_file(env_file_path: str):
+    env_vars = dict()
+
+    with open(env_file_path) as f:
+        lines = f.readlines()
+        for line in lines:
+            key, value = line.split("=")
+            env_vars[key] = value
+
+        f.close()
+
+    return env_vars
+
+
 def create_mod(mod_name: str):
     mod_dir = f"./Mods/{mod_name}"
     scripts_dir = f"{mod_dir}/Scripts"
@@ -20,6 +34,14 @@ def create_mod(mod_name: str):
 
 
 if __name__ == "__main__":
+    env = read_env_file("./.env")
+
+    if "PALWORLD_CLIENT_DIR" not in env and "PALWORLD_SERVER_DIR" not in env:
+        print(
+            "No environment variable set for PALWORLD_CLIENT_DIR or PALWORLD_SERVER_DIR! One of these must be set"
+        )
+        exit(1)
+
     parser = argparse.ArgumentParser(
         prog="Palworld Mod Manager",
         description="Manages mods in this repo for Palworld",
@@ -38,12 +60,6 @@ if __name__ == "__main__":
         action=argparse.BooleanOptionalAction,
         help="Syncs the mods in this repository with the mods folder in the Palworld directory",
     )
-    parser.add_argument(
-        "-d",
-        "--directory",
-        required=True,
-        help="The location where Palworld is installed",
-    )
 
     args = parser.parse_args()
 
@@ -51,22 +67,62 @@ if __name__ == "__main__":
         print("No command given! Use help to see available commands")
         exit(1)
 
-    palworld_dir = args.directory
     curr_dir = os.getcwd()
 
     if args.create:
         create_mod(args.create)
 
     if args.sync:
-        palworld_mods_dir = Path(f"{palworld_dir}/Pal/Binaries/Win64/Mods")
+        if "PALWORLD_CLIENT_DIR" in env:
+            dir = env["PALWORLD_CLIENT_DIR"]
+            client_dir = Path(f"{dir}/Pal/Binaries/Win64/Mods")
+            for mod in client_dir.iterdir():
+                if mod.is_symlink():
+                    mod.rmdir()
+
+        if "PALWORLD_SERVER_DIR" in env:
+            dir = env["PALWORLD_SERVER_DIR"]
+            server_dir = Path(f"{dir}/Pal/Binaries/Win64/Mods")
+            for mod in server_dir.iterdir():
+                if mod.is_symlink():
+                    mod.rmdir()
+
         repo_mods_dir = Path(f"{curr_dir}/Mods")
 
-        for mod in palworld_mods_dir.iterdir():
-            if mod.is_symlink():
-                mod.rmdir()
-
         for mod in repo_mods_dir.iterdir():
-            symlink_dir = Path(f"{palworld_mods_dir.resolve()}/{mod.name}")
-            symlink_dir.symlink_to(mod)
+            enabled_txt = Path(f"{mod.resolve()}/enabled.txt")
+            if not enabled_txt.exists():
+                continue
+
+            with open(enabled_txt.resolve()) as f:
+                lines = f.readlines()
+                is_server_mod = False
+                for line in lines:
+                    key, value = line.split("=")
+                    if key == "SERVER":
+                        is_server_mod = True if value == "true" else False
+
+                if is_server_mod:
+                    if "PALWORLD_SERVER_DIR" not in env:
+                        print(
+                            f"Cannot sink {mod.name}. Palworld server directory not specified"
+                        )
+                        continue
+                    server_dir = env["PALWORLD_SERVER_DIR"]
+                    symlink_dir = Path(f"{server_dir}/Pal/Binaries/Win64/Mods/{mod.name}")
+                    symlink_dir.symlink_to(mod)
+                    print(f"Synced {mod.name} to Palworld server directory")
+                else:
+                    if "PALWORLD_CLIENT_DIR" not in env:
+                        print(
+                            f"Cannot sink {mod.name}. Palworld client directory not specified"
+                        )
+                        continue
+
+                    client_dir = env["PALWORLD_CLIENT_DIR"]
+                    symlink_dir = Path(f"{client_dir}/Pal/Binaries/Win64/Mods/{mod.name}")
+                    symlink_dir.symlink_to(mod)
+                    print(f"Synced {mod.name} to Palworld client directory")
+                f.close()
 
     exit(0)
